# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Prosthuti API",
    "version": "1.0.0",
    "description": "Local Laravel API. Cookie sessions; mutations require CSRF. See API.md for payloads and workflow constraints. Response schemas are intentionally not exhaustive."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "prosthuti_session"
      }
    }
  },
  "paths": {
...
    "/me": {
      "get": {
        "operationId": "ProfileController_show_get",
        "tags": [
          "me"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful JSON response; see integration guide and controller contract."
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Account or role not allowed"
          },
          "404": {
            "description": "Resource unavailable or not owned"
          },
          "422": {
            "description": "Validation error; JSON message and errors"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "parameters": []
      },
      "put": {
        "operationId": "ProfileController_update_put",
        "tags": [
          "me"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful JSON response; see integration guide and controller contract."
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Account or role not allowed"
          },
          "404": {
            "description": "Resource unavailable or not owned"
          },
          "422": {
            "description": "Validation error; JSON message and errors"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "419": {
            "description": "Missing or expired CSRF token/session"
          },
          "409": {
            "description": "State or foreign-key conflict"
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
  }
}